backend/requirements.txt
fastapi==0.109.2
uvicorn==0.29.0
httpx==0.27.0
redis==5.0.1
pydantic==2.6.4
python-dotenv==1.0.1
python-multipart